requirements.txt
Flask==2.1.2
Pillow==9.1.1
rembg==2.0.25
